+++ NEWS.md
@@ -1,5 +1,6 @@
 ## Version `v4.0.0` (in progress)
 ### Changes:
+* Normalize cron notation to the Quartz format (MODSCHED-33)
 ---
 
 ## Version `v3.0.0` (12.03.2025)